Fix WARNING EventTuple:: Tuple 'EventTuple' 'unsigned long' has different sizes on 32/64 bit systems. Casting 'EventInSequence' to 'unsigned long long'
git-svn-id: http://svn.cern.ch/guest/lhcb/Analysis/trunk@164045 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: c213a3f5cb7f8fd6360046d034c9318fff4fefe0

+++ Analysis/Phys/DecayTreeTupleBase/src/component/EventTuple.cpp
@@ -103,7 +103,7 @@ StatusCode EventTuple::execute()
     }
   }
 
-  sc = tuple->column( "EventInSequence", counter("Event").nEntries() );
+  sc = tuple->column( "EventInSequence", (unsigned long long)counter("Event").nEntries() );
   if ( sc.isFailure() ) return sc;
 
   for ( std::vector<IEventTupleTool*>::iterator i = m_tools.begin();
